Use data window instead of all data ever
Before this commit, each session would accumulate data from when
it started running to whenever it stopped; in the meantime, the
memory consumption would grow and grow, and all of the packages
would be pushed to the browser as a bubble chart which would
cause the browser to crash after a while.

This commit changes the way the data is processed so that only
the last 5 minutes of data are considered, except for the Total
Downloads and Unique Users. For Total Downloads, it's easy to
track that accurately for all time just by incrementing a number
(at least until it overflows), and for Unique Users we use a
bloom filter to probabilistically track user counts with a fixed
amount of memory.